+++ src/clojure_lsp/handlers.clj
@@ -142,6 +142,7 @@
   (let [db @db/db]
     {:project-root-uri (:project-root-uri db)
      :project-settings (:project-settings db)
+     :classpath-configs (:classpath-configs db)
      :client-settings (:client-settings db)
      :final-settings (:settings db)
      :port (or (:port db)