+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   - Bump clj-kondo to `2021.09.25` fixing false-positives with potemkin import-var analysis.
   - Re-scan whole project if any clj-kondo config changed for better consistence. #331
   - Fix clojure-lsp not initializing when empty `project.clj`. #579
+  - Support finding config in classpath via new setting`:classpath-config-paths ["my-org/my-lib"]`. #580
 
 - Editor
   - Fix `resolve macro as` code action after regression introduced recently.